requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
playwright==1.48.0
pluggy==1.5.0
pyee==12.0.0
pytest==8.3.3
pytest-anyio==0.0.0
pytest-asyncio==0.24.0
pytest-base-url==2.1.0
pytest-playwright==0.5.2
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
sniffio==1.3.1
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.3
